+++ src/Client/ClientCore/Events/Details/KeyEventDetails.cs
@@ -18,7 +18,7 @@
 using SDL2;
 using Sovereign.EngineCore.Events;
 
-namespace Sovereign.ClientCore.Events;
+namespace Sovereign.ClientCore.Events.Details;
 
 /// <summary>
 ///     Event details associated with a keyboard event.