+++ src/Client/ClientCore/Events/SDLEventAdapter.cs
@@ -17,6 +17,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 using SDL2;
+using Sovereign.ClientCore.Events.Details;
 using Sovereign.ClientCore.Rendering.Gui;
 using Sovereign.EngineCore.Events;
 
@@ -71,26 +72,60 @@ public class SDLEventAdapter : IEventAdapter
                 case SDL.SDL_EventType.SDL_KEYUP:
                     ev = AdaptSdlKeyUp(sdlEv);
                     break;
+
+                case SDL.SDL_EventType.SDL_MOUSEMOTION:
+                    ev = AdaptSdlMouseMotion(sdlEv);
+                    break;
             }
         }
 
         return ev != null;
     }
 
+    /// <summary>
+    ///     Adapts an SDL_QUIT event.
+    /// </summary>
+    /// <param name="sdlEv">SDL event.</param>
+    /// <returns>Internal event.</returns>
     private Event AdaptSdlQuit(SDL.SDL_Event sdlEv)
     {
         return new Event(EventId.Core_Quit);
     }
 
+    /// <summary>
+    ///     Adapts an SDL_KEYDOWN event.
+    /// </summary>
+    /// <param name="sdlEv">SDL event.</param>
+    /// <returns>Internal event.</returns>
     private Event AdaptSdlKeyDown(SDL.SDL_Event sdlEv)
     {
         var details = new KeyEventDetails { Key = sdlEv.key.keysym.sym };
         return new Event(EventId.Client_Input_KeyDown, details);
     }
 
+    /// <summary>
+    ///     Adapts an SDL_KEYUP event.
+    /// </summary>
+    /// <param name="sdlEv">SDL event.</param>
+    /// <returns>Internal event.</returns>
     private Event AdaptSdlKeyUp(SDL.SDL_Event sdlEv)
     {
         var details = new KeyEventDetails { Key = sdlEv.key.keysym.sym };
         return new Event(EventId.Client_Input_KeyUp, details);
+    }
+
+    /// <summary>
+    ///     Adapts an SDL_MOUSEMOTION event.
+    /// </summary>
+    /// <param name="sdlEv">SDL event.</param>
+    /// <returns>Internal event.</returns>
+    private Event AdaptSdlMouseMotion(SDL.SDL_Event sdlEv)
+    {
+        var details = new MouseMotionEventDetails
+        {
+            X = sdlEv.motion.x,
+            Y = sdlEv.motion.y
+        };
+        return new Event(EventId.Client_Input_MouseMotion, details);
     }
 }